Lua improvements (and some fixes) (#10)
* Removed old folder

* Fixed repeated calls to lua function from godot

* gh actions; Ignore errors on `gh release delete`

* Update typings

* luajit; Harden `__index`, only return if method exists for string indexing

* luajit; Add basic Dictionary support

* Use specific tag for godot-sandbox instead of main
Fixes broken builds due to godot-sandbox updates and not updated for TNI

* Updated lua typing

* luajit; Added `Callable.create(f)`

* Added all events to globals.lua

* luajit; Add new callbacks

* Update typing

* Remove old lua typing directory

* Bump godot-sandbox version

* Added dictionary typings

* luajit; Fix Array & Dictionary __index return nil on invalid method names (was always nil form prior attempt)

* Added `tools/.gitignore`

* Update typing

+++ programs/cpp-template/entry.cpp
@@ -103,7 +103,7 @@ static Variant on_game_host_eod() {
 static Variant on_device_spawned(DeviceUnit device) {
 	fetch_mod_api();
     // this is called when a device is spawned.
-	if ((int64_t)device.device_hardware_class() == DeviceUnit::DeviceHardwareClass::NETWORK_SWITCH) {
+	if ((int64_t)device.device_hardware_class() == (int64_t)DeviceUnit::DeviceHardwareClass::NETWORK_SWITCH) {
 		LogicController lc = device.logic_controller();
 		int64_t original_nbw = lc.installed_nbw();
 		int64_t new_nbw = original_nbw * 2;
